requirements.txt
annotated-types==0.7.0
anyio==4.4.0
boto3==1.34.161
botocore==1.34.161
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
fastapi==0.112.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
jmespath==1.0.1
passlib==1.7.4
psycopg2==2.9.9
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6